+++ ghost/admin/app/templates/posts.hbs
@@ -27,52 +27,32 @@
     </GhCanvasHeader>
 
     <section class="view-container content-list">
-        <div class="{{if this.feature.memberAttribution 'gh-list-sticky'}}">
-            <ol class="posts-list gh-list {{unless this.postsInfinityModel "no-posts"}} {{if this.feature.memberAttribution 'feature-memberAttribution'}}">
-                {{#if this.postsInfinityModel}}
-                    <li class="gh-list-row header">
-                        <div class="gh-list-header gh-posts-title-header">Title</div>
+        <ol class="posts-list gh-list {{unless this.postsInfinityModel "no-posts"}} feature-memberAttribution">
 
-                        {{#if (and this.feature.memberAttribution (not this.session.user.isContributor)) }}
-                        <div class="gh-list-header gh-posts-status-header"></div>
-                            <div class="gh-list-header gh-posts-signups-header">Signups</div>
-                            <div class="gh-list-header gh-posts-conversions-header">Paid</div>
-                        {{/if}}
-                        {{#if (and (not-eq this.settings.membersSignupAccess "none") (not-eq this.settings.editorDefaultEmailRecipients "disabled") (not this.session.user.isContributor) this.feature.emailAnalytics)}}
-                            <div class="gh-list-header gh-posts-sends-header">Sends</div>
-                            <div class="gh-list-header gh-posts-opens-header">Opens</div>
-                        {{/if}}
-                        {{#if (or (not this.feature.memberAttribution) this.session.user.isContributor) }}
-                            <div class="gh-list-header gh-posts-status-header">Status</div>
-                        {{/if}}
-                    </li>
-                {{/if}}
-
-                {{#each this.postsInfinityModel as |post|}}
-                    <PostsList::ListItem
-                        @post={{post}}
-                        data-test-post-id={{post.id}}
-                    />
-                {{else}}
-                <li class="no-posts-box">
-                    <div class="no-posts">
-                        {{#if this.showingAll}}
-                            {{svg-jar "posts-placeholder" class="gh-posts-placeholder"}}
-                            <h4>Start creating content.</h4>
-                            <LinkTo @route="editor.new" @model="post" class="gh-btn gh-btn-green">
-                                <span>Write a new post</span>
-                            </LinkTo>
-                        {{else}}
-                            <h4>No posts match the current filter</h4>
-                            <LinkTo @route="posts" @query={{hash type=null author=null tag=null}} class="gh-btn">
-                                <span>Show all posts</span>
-                            </LinkTo>
-                        {{/if}}
-                    </div>
-                </li>
-                {{/each}}
-            </ol>
-        </div>
+            {{#each this.postsInfinityModel as |post|}}
+                <PostsList::ListItem
+                    @post={{post}}
+                    data-test-post-id={{post.id}}
+                />
+            {{else}}
+            <li class="no-posts-box">
+                <div class="no-posts">
+                    {{#if this.showingAll}}
+                        {{svg-jar "posts-placeholder" class="gh-posts-placeholder"}}
+                        <h4>Start creating content.</h4>
+                        <LinkTo @route="editor.new" @model="post" class="gh-btn gh-btn-green">
+                            <span>Write a new post</span>
+                        </LinkTo>
+                    {{else}}
+                        <h4>No posts match the current filter</h4>
+                        <LinkTo @route="posts" @query={{hash type=null author=null tag=null}} class="gh-btn">
+                            <span>Show all posts</span>
+                        </LinkTo>
+                    {{/if}}
+                </div>
+            </li>
+            {{/each}}
+        </ol>
 
         <GhInfinityLoader
             @infinityModel={{this.postsInfinityModel}}
